# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Issue Tracker - API
  version: 1.0.0
  description: OpenAPI-like summary for the local AI Issue Tracker API routes
servers:
  - url: http://localhost:3000
paths:
...
  /api/issues:
    get:
      summary: List all issues
      responses:
        '200':
          description: Array of issues
          content:
            application/json: {}
    post:
      summary: Create a new issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueCreate'
      responses:
        '200':
          description: Created
...
  /api/analytics:
    get:
      summary: Get analytics and trends
      parameters:
        - name: timeRange
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
      responses:
        '200':
          description: Analytics payload
...
  /api/ai/sops/{id}/export:
    post:
      summary: Export SOP in specified format
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                format:
                  type: string
                  enum: [markdown, txt, pdf]
      responses:
        '200':
          description: Export payload { content, filename }
...
components:
  schemas:
...
    IssueCreate:
      type: object
      required: [title, description, category, severity, environment, systemAffected, reporterName, reporterEmail]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        severity:
          type: string
        environment:
          type: string
        systemAffected:
          type: string
        reporterName:
          type: string
        reporterEmail:
          type: string
...