# src/app/api/ai/sops/[id]/export/route.ts
import { NextRequest, NextResponse } from 'next/server'
import { db } from '@/lib/db'

export async function POST(
  request: NextRequest,
  { params }: { params: { id: string } }
) {
  try {
    const sopId = params.id
    const body = await request.json()
    const { format } = body

    // Get the SOP details
    const sop = await db.aiSop.findUnique({
      where: { id: sopId },
      include: {
        issue: {
          select: {
            title: true,
            category: true,
            severity: true
          }
        }
      }
    })

    if (!sop) {
      return NextResponse.json(
        { error: 'SOP not found' },
        { status: 404 }
      )
    }

    let content = ''
    const filename = `SOP_${sop.issue.title.replace(/[^a-zA-Z0-9]/g, '_')}`

    if (format === 'markdown') {
      content = `# Standard Operating Procedure

**Issue:** ${sop.issue.title}
**Category:** ${sop.issue.category}
**Severity:** ${sop.issue.severity}
**Generated:** ${sop.generatedAt.toLocaleString()}

---

${sop.sopText}

---

*This SOP was automatically generated by AI Issue Tracker based on resolved issues.*`
    } else if (format === 'txt') {
      content = `STANDARD OPERATING PROCEDURE
=====================================

Issue: ${sop.issue.title}
Category: ${sop.issue.category}
Severity: ${sop.issue.severity}
Generated: ${sop.generatedAt.toLocaleString()}

${sop.sopText}

---
This SOP was automatically generated by AI Issue Tracker based on resolved issues.`
    } else if (format === 'pdf') {
      // For PDF, we'll return markdown content and let the frontend handle PDF generation
      content = `# Standard Operating Procedure

**Issue:** ${sop.issue.title}
**Category:** ${sop.issue.category}
**Severity:** ${sop.issue.severity}
**Generated:** ${sop.generatedAt.toLocaleString()}

---

${sop.sopText}

---

*This SOP was automatically generated by AI Issue Tracker based on resolved issues.*`
    } else {
      return NextResponse.json(
        { error: 'Unsupported format' },
        { status: 400 }
      )
    }

    return NextResponse.json({
      content,
      filename: `${filename}.${format}`
    })
  } catch (error) {
    console.error('Failed to export SOP:', error)
    return NextResponse.json(
      { error: 'Failed to export SOP' },
      { status: 500 }
    )
  }
}